+++ src/template.php
@@ -1,17 +1,21 @@
 <?php
 header("Content-type: text/xml");
-echo '<?xml version="1.0" encoding="utf-8"?>';
+echo '<?xml version="' . $this->options['version'] . '" encoding="' . $this->options['encoding'] . '"?>';
 ?>
 
-<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
+<urlset xmlns="<?= $this->options['xmlns']; ?>">
 <?php
 if( ! empty( $urls ) ) :
   foreach( $urls as $row ) :
     if( ! empty( $row ) ) : ?>
   <url>
-<?php foreach( $row as $key => $value ) : ?>
-    <<?= $key; ?>><?= $value; ?></<?= $key; ?>>
-<?php endforeach; ?>
+<?php
+  foreach( $row as $key => $value ) {
+    if($value) {
+      echo "\t<$key>$value</$key>\n";
+    }
+  }
+?>
   </url>
 <?php
 endif;