+++ changelog.md
@@ -1,3 +1,8 @@
+# 1.1
+
+- Change defaults from default values to no values.
+- Added new options and defaults for version, encoding and xmlns.
+
 # 1.0
 
 - Initial release